Deep Archaeology Patch
Deep Archaeology checks if one of two FW missions are offered to see if you know the Pug. In ScS, you don't have to meet the Pug, but you can, and if you do so, the campaign now sets the FW reconciliation one as offered so Deep Archaeology works properly.